GSC QUERY COLLECTOR -2/requirements.txt
google-auth>=2.0.0
google-api-python-client>=2.0.0
pyyaml>=6.0
flask>=3.0.0
